#6697 delete user api update

+++ src/main/java/edu/harvard/iq/dataverse/api/Admin.java
@@ -324,16 +324,21 @@ public class Admin extends AbstractApiBean {
 		return error(Response.Status.BAD_REQUEST, "User " + identifier + " not found.");
 	}
 
-	@DELETE
-	@Path("authenticatedUsers/{identifier}/")
-	public Response deleteAuthenticatedUser(@PathParam("identifier") String identifier) {
-		AuthenticatedUser user = authSvc.getAuthenticatedUser(identifier);
-		if (user != null) {
-			authSvc.deleteAuthenticatedUser(user.getId());
-			return ok("AuthenticatedUser " + identifier + " deleted. ");
-		}
-		return error(Response.Status.BAD_REQUEST, "User " + identifier + " not found.");
-	}
+    @DELETE
+    @Path("authenticatedUsers/{identifier}/")
+    public Response deleteAuthenticatedUser(@PathParam("identifier") String identifier) {
+        AuthenticatedUser user = authSvc.getAuthenticatedUser(identifier);
+        if (user != null) {
+            
+            if (!dvObjSvc.findByAuthenticatedUserId(user).isEmpty()){
+                return error(Status.BAD_REQUEST, "Could not delete AuthenticatedUser " + identifier + " because the user has created Dataverse object(s). " );
+            }
+            
+            authSvc.deleteAuthenticatedUser(user.getId());
+            return ok("AuthenticatedUser " + identifier + " deleted. ");
+        }
+        return error(Response.Status.BAD_REQUEST, "User " + identifier + " not found.");
+    }
         
         
         
@@ -811,16 +816,20 @@ public class Admin extends AbstractApiBean {
 		return ok(response);
 	}
 
-	@DELETE
-	@Path("authenticatedUsers/id/{id}/")
-	public Response deleteAuthenticatedUserById(@PathParam("id") Long id) {
-		AuthenticatedUser user = authSvc.findByID(id);
-		if (user != null) {
-			authSvc.deleteAuthenticatedUser(user.getId());
-			return ok("AuthenticatedUser " + id + " deleted. ");
-		}
-		return error(Response.Status.BAD_REQUEST, "User " + id + " not found.");
-	}
+    @DELETE
+    @Path("authenticatedUsers/id/{id}/")
+    public Response deleteAuthenticatedUserById(@PathParam("id") Long id) {
+        AuthenticatedUser user = authSvc.findByID(id);
+        if (user != null) {
+            if (!dvObjSvc.findByAuthenticatedUserId(user).isEmpty()) {
+                return error(Status.BAD_REQUEST, "Could not delete AuthenticatedUser " + user.getIdentifier() + " because the user has created Dataverse object(s). ");
+            }
+
+            authSvc.deleteAuthenticatedUser(user.getId());
+            return ok("AuthenticatedUser " + id + " deleted. ");
+        }
+        return error(Response.Status.BAD_REQUEST, "User " + id + " not found.");
+    }
 
 	@Path("roles")
 	@POST
